backend/pyproject.toml
[project]
name = "safezone-ai-backend"
version = "1.0.0"
description = "SafeZone AI — Real-Time Workplace Safety Intelligence Agent"
requires-python = ">=3.11"

dependencies = [
    # Vision Agents (WORKING SET)
    "vision-agents>=0.3.8",
    "vision-agents-plugins-getstream==0.3.8",
    "vision-agents-plugins-gemini==0.3.8",
    "vision-agents-plugins-elevenlabs==0.3.8",
    "vision-agents-plugins-deepgram==0.3.8",
    "vision-agents-plugins-smart-turn==0.3.8",
    "vision-agents-plugins-ultralytics==0.1.6",
    "ultralytics>=8.2.0",
    "anthropic>=0.40.0",
    "getstream>=0.3.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
